chore(prediction): Avoid create new store in leaderboard
