+++ src/pages/prediction/leaderboard.tsx
@@ -1,18 +1,9 @@
-import LocalReduxProvider from 'contexts/LocalRedux/Provider'
-import makeStore from 'contexts/LocalRedux/makeStore'
-import reducers, { initialState } from 'state/predictions'
-import ConfigProvider, { config } from '../../views/Predictions/context/ConfigProvider'
+import BnbConfigProvider from '../../views/Predictions/context/BnbConfigProvider'
 
 import PredictionsLeaderboard from '../../views/Predictions/Leaderboard'
 
-const store = makeStore(reducers, initialState, config)
-
-export default () => {
-  return (
-    <ConfigProvider>
-      <LocalReduxProvider store={store}>
-        <PredictionsLeaderboard />
-      </LocalReduxProvider>
-    </ConfigProvider>
-  )
+export default function Leaderboard() {
+  return <PredictionsLeaderboard />
 }
+
+Leaderboard.Layout = BnbConfigProvider
